refactor: import hooks directly
